Start the GitForWindowsHelper GitHub App project
This GitHub App is "already in use" by
https://github.com/git-for-windows/git-for-windows-automation/, where
its credentials allow us to "mirror" a Check Run from that repository,
where the GitHub workflow actually runs, to another repository, where
the code lives that is being deployed.

To register the GitHub App, we needed a URL to which to deliver the
webhook events. The URL we used pointed to a relatively simple Azure
Function that was implemented specifically for that purpose.

Since we already have an Azure Function that receives the webhook events
intended for the registered GitHub App, we might just as well make use
of it.

The original Azure Function was very, very simple and did not do much
more than say "Ohai!".

Out of that Azure Function, though, arose
https://github.com/dscho/leave-me-alone-github-app, which is slightly
more than a toy: It is an excellent start for implementing any minimal
GitHub App. And we will make use of that excellent start in the next
commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>