Add a node.js script to test webhook delivery locally
With this script, developers can copy/paste the headers and the payload
from https://github.com/settings/apps/gitforwindowshelper/advanced, save
the result in a file, and then simulate the re-delivery by running this
script with the file as command-line argument.

If anything interesting needs to be tested (such as sending GitHub API
requests using the installation token of the GitHub App), the
`local.settings.json` file needs to be set up like it would have to be
for running the Azure Function locally, as described in the `README.md`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>