Add functions to determine collaborators' permission
We will want to ensure that only users with write permissions can
trigger certain slash commands. To that end, let's add a function to
determine the permissions.

This is a bit more tricky than previously-added functions because there
is no REST API call to determine that. Only GraphQL, which makes things
infinitely more complex.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>