Add a function to dispatch workflow runs
One of the main building blocks of the "new" Git for Windows package
deployment process is to let GitHub workflows do most of the job,
triggered by the GitForWindowsHelper GitHub App.

To this end, the workflows accept a `workflow_dispatch` trigger and this
here function allows to start those workflow runs.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>