https-request: avoid to always log failures
Some requests are intended to _potentially_ fail, e.g. when trying to
determine the identity associated with the current token (trying to get
the user associated with an installation token will fail, and we want to
fall back to get the associated app in that case instead).

Let's instead report the method and the path in the response (but not
the headers, which contain secrets).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>