slash-commands: optionally trigger the Git Artifacts pipeline
In Git for Windows, releases are done by first rebasing Git for Windows'
branch thicket on top of the new Git version, then opening a Pull
Request, and after the CI has passed successfully, trigger an Azure
Pipeline called "Git Artifacts" that uses that PR branch to generate all
the artifacts required for a release.

Due to technical reasons, it is relatively easy to trigger this Azure
Pipeline with inconsistent/incorrect parameters.

Let's avoid that by offering the `/git-artifacts` command to do all
that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>