/deploy: do queue check-run
The neat trick of the `build-and-deploy` workflow is that it "mirrors" a
check-run into the target repository.

But that check-run is currently only mirrored once the workflow runs and
executes the respective step, which necessarily takes a few seconds
because the workflow first has to check out the code to run the
deployment.

Let's let the GitHub App already add the check-run, in `queued` state,
to provide a more immediate visual feedback to the `/deploy` command.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>