guessComponentUpdateDetails: get version from content on aggregated issues
Aggregated issues have titles like "[New bash version] 3 new items". They don't
contain a version number, but we can get the titles of the original issues that
where aggregated into one from headlines in the issue content. The first one of
them is the latest one, which is usually the version we care about.

Signed-off-by: Matthias Aßhauer <mha1993@live.de>