/deploy: populate details URL
GitHubs check runs provide a neat little "view more details" link (see for example [1])
that should allow users to view the status of the check run (beyond running/queued/success/failure).
Instead of just statically linking to [2], populate the URL to link to the workflow run that
we're waiting for.

[1] https://github.com/git-for-windows/build-extra/pull/451/checks?check_run_id=10110786774
[2] https://github.com/git-for-windows/gfw-helper-github-app

Signed-off-by: Matthias Aßhauer <mha1993@live.de>