/git-artifacts: do not require a "Rebase to <version>" title
Today, I want to release Git for Windows v2.39.0(2), a version for which
no upstream Git version was released (v2.39.1 is still a few weeks out:
https://lore.kernel.org/git/xmqqcz8csz5k.fsf@gitster.g/).

So I created a Pull Request with a minor change that I wanted to do
anyway, as an excuse for the Git artifacts build to run.

Let's allow that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>