/deploy: optionally accept a package name as argument
Sometimes the name of the package cannot be determined from the Pull
Request, or we need to deploy a dependent package in addition to the one
that is changed in that Pull Request.

In either case it is necessary to specify explictly which package to
deploy. Let's support `/deploy <package-name>` for these instances.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>