/deploy: use separate check runs for each architecture
For MSys packages we start two build workflows, that can succeed or fail
independently from one another. Representing their combined outcome in one
check run, whose result only depends on the first workflow to finish is not
great. Using one check run for each workflow makes it easier to see which
workflow failed when one of them fails and allows us to see at a glance if
one of the workflows is still running after the other one completed.

Signed-off-by: Matthias Aßhauer <mha1993@live.de>