/deploy: trigger clang-aarch64 builds when appropriate
Since we now have support to automatically spin up/tear down
Windows/ARM64 runners as needed, we will want to make use of that by
deploying not only the i686/x86_64 MINGW packages, but by deploying the
aarch64 packages, too.

Note that the `git-credential-manager`, `git-lfs` and `wintoast`
packages are special: Git for Windows does not actually build the ARM64
variants of these packages via clang, but instead either downloads the
artifacts from the respective upstream projects or builds using Visual
Studio (which supports cross-compiling on hosted agents, in contrast to
MSYS2's clang). Therefore, these packages do _not_ require a separate
workflow run for ARM64.

An alternative to special-casing these packages in the
GitForWindowsHelper GitHub App would have been to turn the
`build-and-deploy` workflow into a dynamic matrix that spins up a
separate job for the `aarch64` architecture if, and when, needed.
However, that would have made that workflow even more complex than it
already is, and therefore that alternative was rejected.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>