Add a test script to help debugging `/open pr`
The recent attempt at opening PRs for OpenSSL at
https://github.com/git-for-windows/git/issues/4273#issuecomment-1422172149
failed to produce the correct `version` parameter: instead of the
desired `1.1.1t` it came up with `1.1_1t` (i.e. only the first
underscore was converted to a dot).

To help with debugging such issues, let's add a script that allows for
reproducing the functionality locally:

	./test-open-pr.js <url-to-issue>

will output the package name and the version that was determined from
that ticket.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>