guessReleaseNotes: be smarter when looking at PRs
The default for PRs is to say "This closes <url>" when the `open-pr`
workflow creates a PR, referencing the original issue.

So it would not really help if we linked to that issue in the release
notes as much as if we looked for a URL in that referenced issue.

Therefore, let's follow that link and look for such a better URL.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>