Automatically trigger the `git-artifacts` runs upon a completed `tag-git` run
This commit is part of Git for Windows' transition off of Azure
Pipelines onto GitHub workflows.

One thing that was changed during that transition is that the tasks were
split up into several workflow runs instead of having one ginormous `Git
Artifacts` Pipeline that does everything from tagging the Git version to
building the Pacman packages in matrix builds, to building the
installers/portable Gits/MinGits/etc in another set of matrix builds, to
generating the release notes as well as the announcement mail.

Now, the Git version is tagged in the `tag-git` workflow, and that
workflow also generates the release notes and announcement mail
template (the SHA-256s of the generated artifacts are part of both,
therefore the final version can only be constructed once those artifacts
are available).

Once the `tag-git` workflow has run successfully, the `git-artifacts`
workflow needs to be started multiple times, once per supported
architecture (currently x86_64 and i686, hopefully soon also aarch64).

This commit teaches GitForWindowsHelper to do that by reacting when a
`tag-git` run completed, triggering the `git-artifacts` runs.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>