Teach `/release` to trigger the `release-git` workflow
This concludes the transition of Git for Windows' release process off of
Azure Pipelines onto GitHub workflows.

The added code is a bit more verbose than the previous one because we
have to do all the hard work of maintaining the Check Run in the
git-for-windows/git PR manually.

This is actually a net improvement over the previous situation, where
there was _no_ update in that PR whatsoever as to the current state of
the release.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>