Fix updating the queued Check Run in case of failure
When something goes wrong, we want to update the Check Run with enough
information to help readers figure out how to fix it.

However, we forgot to specify `title` and `summary` and therefore that
GitHub REST API call to update the Check Run failed, hiding the original
failure.

Let's not forget titles and summaries.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>