/release: pass the git-artifacts _workflow_ runs
Since we mirror Check Runs into a different repository, it _can_ be
quite confusing which ID is required. But for the `release-git` workflow
it is clear: it needs the artifacts which are attached to the
`git-artifacts` workflow runs.

But we have the Check Run only.

But then, the `output.text` attribute has a link to the workflow, from
which we can easily determine the workflow run ID.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>