tests: verify that GCM deployments trigger only one combined run
Git Credential Manager is special in that we do not actually _build_ the
program, we simply re-bundle whatever the project published at
https://github.com/git-ecosystem/git-credential-manager.

Therefore, we can get away with _not_ triggering a separate aarch64
run, but we can bundle i686, x86_64 and aarch64 packages in a single
workflow run.

Verify that this is actually what happens.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>