DRY up the `/deploy` handling
There was a lot of copy/edited code in that part, owing to the fact that
it is not _completely_ trivial to generalize adding the intended Check
Runs, triggering the intended workflow runs, updating the PR comment
with said information, and then updating the Check Runs, too.

Unfortunately, there is no really good way to present this refactoring,
as it collapses the separate MSYS2/MINGW code paths into a single one.

The end result looks a lot cleaner, though, with the only tricky bit
being how the comment is constructed (since there is now a variable
number of workflow runs that may need to be listed).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>