/deploy: no longer build the MSYS2 runtime packages for both architectures
The purpose of introducing a separate `msys2-runtime-3.3` package is to
allow supporting Git for Windows on i686 setups, even though it cannot
receive updates to the latest MSYS2 runtime versions any longer because
the MSYS2 project dropped i686 support a long time ago.

To that end, let's actually build `msys2-runtime-3.3` _only_ for i686,
and `msys2-runtime` _only_ for x86_64.

That split will allow us to change the `msys2-runtime-3.3` package
definition so that it supersedes ("replaces") `msys2-runtime` and
automagically updates the 32-bit Git for Windows SDKs out there.

This is a step toward phasing out i686 support of Git for Windows, see:
https://github.com/git-for-windows/git/issues/4279#issue-1577622335 for
more details about that plan.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>