Add a function to determine whether a package has been deployed completely
In https://github.com/git-for-windows/git/issues/4567 it was pointed out
that Git LFS v3.3.0 is shipped with Git for Windows v2.42.0, but the
release notes state that the current version is v3.4.0.

In preparation for making `/add relnote` a bit safer to use, introduce
the functionality in the GitForWindowsHelper GitHub App to determine
whether a given version of a given package has been deployed, and if
not, which parts are missing (indicated by the respective Pacman
repository URL(s)).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>