/add relnote: verify that the package was deployed correctly
When calling `/add relnote` from a `component-updates` issue (and
letting the GitForWindowsHelper GitHub App figure out the package name
and version from the title) or from a corresponding PR, we would like to
ensure that the corresponding package files have been deployed
correctly.

In https://github.com/git-for-windows/git/issues/4523, we forgot to
deploy all the package files, and it's an easy mistake to make, hence
this commit, to help prevent the same in the future).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>