/add relnote: respect MinTTY's "epoch"
In the Pacman universe, there is the concept of an "epoch", which is
part of the versioning scheme and which has more weight than the
`pkgver` value.

This concept allows following projects which "re-set" their version, for
example (e.g. switching from a date-based "version" to an actual version
that would naturally sort way below the previous date-based one).

In the Git for Windows project, we incremented the epoch of the `mintty`
package, originally simply to override MSYS2's `mintty`, even if that
was technically not even necessary.

Yet it is part of Git for Windows' legacy that the `/add release note`
command needs to keep in mind when trying to verify that the `mintty`
packages have been deployed properly.

Thankfully, `mintty` is the only package built by Git for Windows that
has an epoch set, and hence it is the only package that needs to be
treated specially in this way.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>