guessReleaseNotes(msys2-runtime): look harder for Cygwin's release notes
Cygwin's release notes link on their home page is sometimes not updated
immediately after a new version has been made available.

In the past, i.e. before Git for Windows built and deployed MSYS2
runtime versions via GitHub Actions, there was a fall-back to look for
that link directly in the mailing list archives.

In the meantime, Cygwin offers access to their mailing list archive via
a public-inbox system, which makes the search even nicer.

So let's do this, add a fall-back to look for the release notes directly
on the mailing list.

This corresponds to the code at
https://github.com/git-for-windows/build-extra/blob/356779ddb189/please.sh#L1731-L1735

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>