/git-artifacts: add the `git-artifacts` runs to the PR comment more reliably
When `/git-artifacts` has to trigger a `tag-git` run first, the actual
`git-artifacts-<arch>` jobs are triggered as "cascading runs". At that
stage, the connection to the original PR comment ID is lost, and
therefore that comment is no longer updated to link to those
`git-artifacts-<arch>` runs.

With this patch, GitForWindowsHelper learns to find that PR comment ID
when a completed `tag-git` check run triggers the `git-artifacts-` runs
and then to update that comment accordingly.

This addresses https://github.com/git-for-windows/gfw-helper-github-app/issues/38

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>