Add a script to remove unwanted installations
Every once in a while, somebody installs the GitForWindowsHelper GitHub
App on their repository.

Seeing as we have checks that ensure that only webhook events from the
intended repositories are handled, this does not affect those
repositories at all.

However, it does send a ton of webhook events to those repositories, and
when I go looking in
https://github.com/organizations/git-for-windows/settings/apps/gitforwindowshelper/advanced
for any specific delivery (e.g. to see the response from the Azure
Function App in case something went wrong), it is incredibly tedious to
click "Load more deliveries" a thousand times until I finally get to the
interesting webhook event delivery.

Let's avoid that by simply running (manually) this script from time to
time, to uninstall the App on unintended repositories, to avoid those
unnecessary and unwanted webhook deliveries.

This script requires the same `local.settings.json` file as is needed
for testing the Azure Function locally (for details, refer to
`README.md#run-the-azure-function-locally`); That is where it gets the
credentials required to perform the uninstallations.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>