Add a convenience script to find specific webhook deliveries
In the previous commit, I added a script to delete existing
installations of the GitForWindowsHelper GitHub App for unintended
repositories, to avoid cluttering the list of webhook deliveries.

I ran this script a couple times in the past, but it was always a bit
late for that because I had to click literally over 200 times on that
button just to get to payloads that were delivered two days prior.

A better way is to use the GitHub REST API to let a script search for
specific deliveries, say, for a `check_run` that has completed around a
given time.

I used this script today to great effect to find the `check_run` and
`workflow_job` webhook payloads that I mentioned in
https://github.com/git-for-windows/git-for-windows-automation/pull/60

It uses a bit of an ad-hoc strategy to avoid having to load sequential
pages of deliveries until the desired time window is reached. That
strategy could probably be improved, but it seems to work well enough
for now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>