Add the `/updpkgsums` slash command
As with most of the other slash commands, this triggers a GitHub
workflow run in `git-for-windows-automation`, in this instance,
unsurprisingly the `upkpkgsums.yml` one.

The idea is to help contributors with their Pull Requests that fail CI
builds solely due to mismatched checksums after modifying, adding or
removing files from a package.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>