/deploy: respect the `i686:` prefix of PR titles
To allow for building i686 packages that are no longer built by the
MSYS2 project, let's establish the paradigm that Pull Requests whose
titles start with `i686:` are to be deployed only to the 32-bit flavor
of the Git for Windows Pacman repository.

Note: Since it is not customary to specifically point out each and every
such i686 build in the release notes, this commit does not take care of
`getMissingDeployments()` nor of the `/add relnote` command.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>