After a successful `release-git` run, close the corresponding PR
In Git for Windows' release process, there are always
`git-for-windows/git` PRs whose tip commit is slated to be tagged.

As part of the release process, after the `/git-artifacts`
slash command has concluded successfully and its artifacts have been
validated manually, the `/release` slash command is issued, which
triggers the `release-git` workflow run.

Once this `release-git` workflow run has succeeded, it is time to close
the PR. However, it must not be merged (which would throw the
merging-rebase automation) but instead the PR's tip revision needs to be
pushed as new tip commit of the `main` branch.

The code added in this commit accomplishes that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>