WIP: Automatically close Git for Windows' milestones after a release
This is a work in progress! The idea is that as part of closing a `[New
git release]` ticket that is associated with a Git for Windows release,
we will close the current milestone and open the next one. The current
work-in-progress script was used to successfully close the v2.44.0
milestone and open the `Next release` one. So what's missing? This:

- The current script needs to be replaced by a proper route in
  `index.js`

- More validations are needed, so that we do not close, say, a milestone
  that does not even have any issues assigned to it. Or a milestone that
  is for another version.

- Tests.

This will address
https://github.com/git-for-windows/gfw-helper-github-app/issues/7.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>