Find the `/git-artifacts` comment even yet more reliably
It _has_ happened in the past that multiple `/git-artifacts` calls were
required until things worked, e.g.
https://github.com/git-for-windows/git/pull/4330#issuecomment-1458893794
which did not work and was followed by
https://github.com/git-for-windows/git/pull/4330#issuecomment-1458894252
which also did not work, and was followed by
https://github.com/git-for-windows/git/pull/4330#issuecomment-1458894252
which pretended to work, but didn't, and was followed by
https://github.com/git-for-windows/git/pull/4330#issuecomment-1458896122
which suceeded, at long last, to produce the Git Artifacts.

So let's make extra sure that we are finding the correct
`/git-artifacts` comment to amend.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>