self-hosted-arm64-runners: specify target repo
The create-azure-self-hosted-runners workflow has an input called `runner_repo` which can be used to specify the repository that the runner will be deployed to. It defaults to git-for-windows-automation, which is not the right one in all cases.

Let's set the `runner_repo` input to the repository that requested the runner.

Ref: https://github.com/git-for-windows/git-for-windows-automation/blob/main/.github/workflows/create-azure-self-hosted-runners.yml#L18C7-L18C18
Signed-off-by: Dennis Ameling <dennis@dennisameling.com>