/open pr: support running in git-for-windows/msys2-runtime
The MSYS2 runtime is a special thing indeed. Most of the time, patches
for Git for Windows' variant of this component come in not via MSYS2,
but via PRs in https://github.com/git-for-windows/msys2-runtime.

I already taught the `open-pr` GitHub workflow about this in
https://github.com/git-for-windows/git-for-windows-automation/pull/87,
and now it is time to extend the `/open pr` slash command to cover this
scenario, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>