guess release notes: special-case cURL (even in MINGW-packages)
In https://github.com/git-for-windows/gfw-helper-github-app/pull/34, I
taught the GitForWindowsHelper GitHub App to generate better release
notes for cURL.

However, I failed to notice that the code changes would work only for
the `component-updates` issues as well as for the MSYS2-packages PRs,
but not for the MINGW-packages PRs because the package name that is
derived from the latter PRs has the `mingw-w64-` prefix (and must have
it, for the "is this already deployed?" check).

Let's make sure that that special-casing (as well for cURL as well as
for OpenSSL) work even in MINGW-packages PRs.

Helped-by: Matthias Aßhauer <mha1993@live.de>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>