vscode(launch): offer to debug reacting to a snapshot build
The idea is that that the `upload-snapshot` workflow is triggered as
soon as the last `git-artifacts-*` check run has completed. Let's make
it easy on myself by taking an existing webhook event that was delivered
and that ideally should have triggered all that, and then develop the
relevant code leaning heavily on VS Code's debugger.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>