When all `git-artifacts` runs are done for a snapshot, upload it
The `tag-git` workflow can be used to handle Git for Windows' snapshots
via the `snapshot` Boolean that is part of its inputs.

The idea is that once that workflow has completed, the `git-artifacts`
workflow runs are triggered automatically via GitForWindowsHelper, one
per supported CPU architecture.

Now this commit introduces the logic to trigger the `upload-snapshot`
workflow once all of those `git-artifacts` runs completed successfully.

Naturally, we have to be careful to do this only when the snapshot has
not yet been uploaded, and only when the snapshot is built from a commit
that is actually on git-for-windows/git's default branch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>