Add support for the `/synchronize-sdks` command
This slash command will trigger the `sync` workflows in the Git SDK
repositories, which updates them according to the Pacman repositories.

This is particularly useful after deploying a package, to make sure that
the next `git-artifacts` build will pick up the new package version.

For convenience, there is also the `/sync` shortcut.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>