test-pr-comment-delivery: have an explicit way to avoid triggering runs
The main reason why some of the settings are disabled in this script is
so that no state is changed on github.com when debugging a webhook
payload event locally.

However, in some cases the GitForWindowsHelper app needs to query
GitHub's REST API using an access token because some queries cannot be
issued without authorization. For example, for some reason or other
reading workflow runs' logs requires a token (which is different than
what Azure Pipelines does, from which GitHub Actions were derived, so
this is doubly curious).

To help with such debugging scenarios, have a separate way to prevent
specifically triggering workflow runs, which would be the worst thing to
do inadvertently during a debugging session.

By default, the private key and Azure Pipelines trigger token are still
disabled in the script, but now it is at least somewhat safe to comment
out the private key disabling so that above-mentioned queries can be
debugged locally.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>