snapshot uploading: fix the ahead/behind by logic
The logic was _exactly_ inverted. What it _did_ want to verify was that
the commit for which the snapshot was built is reachable from `main`.
What it verified instead was that `main`'s tip commit is reachable from
said commit. 🤦

I noticed this only today, when a successful `git-artifacts` run in the
v2.48.1 PR at https://github.com/git-for-windows/git/pull/5411 tried to
upload a snapshot, and the (correct) ahead/behind by logic in the
`upload-artifacts` workflow failed (but then succeeded when I re-ran the
workflow after releasing Git for Windows v2.48.1).

Let's invert the logic so that it does what it is supposed to do.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>