snapshot uploading: do upload the regular Git for Windows versions
A regular Git for Windows version is branch-deployed in a Pull Request:
It is built via the `/git-artifacts` slash command and then published
via the `/release` slash command.

Crucially, this build needs to be validated manually between the time
when the `git-artifacts` workflows are done and the time when they are
uploaded, and during that time window, the artifacts should _not_ be
published as snapshots because if validation fails, they need to be
rebuilt.

The preceding two commits fixed the bug where the `upload-snapshot`
workflow was run in that time window by mistake.

However, when _do_ we want to upload the freshly-built official version
as a new snapshot? Why, of course when the `main` branch is pushed.

Except that we do not push at all, the `/release` slash command just
updates the `main` ref, knowing fully well that no objects need to be
pushed because `main` must fast-forward to the the PR branch when
releasing a new Git for Windows version; This is enforced. And updating
the ref via the REST API does not trigger the `push` event at all.

Therefore we need to manually pretend that a `push` happened, and then
all will be good again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>