upload-snapshot: queue the Check Run using a valid token
The `getToken()` method does not need any parameters when being called
in `slash-commands.js` because there, the `owner`/`repo` values are
obvious. But when triggering an `upload-snapshot` workflow, we are
juggling three different repositories, therefore we absolutely need to
specify for which repository the token needs to target.

This was the _actual_ reason why the `upload-snapshot` run was not
triggered after v2.48.1 was released, the conclusions I drew in
https://github.com/git-for-windows/gfw-helper-github-app/pull/124 were
incorrect.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>