upload-snapshot: fail earlier on invalid `tag-git` Check Runs
When handling a `push` event on `git-for-windows/git`, it is necessary
to validate that the `tag-git` Check Run has the information needed to
determine the `git-artifacts` runs.

If that information is not there (e.g. because `tag-git` was
accidentally updated and now contains invalid/unparseable data), we
should error out earlier.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>