upload-snapshot: fix discovery of the `git-artifacts` runs
When a `push` wants to trigger a new snapshot to be uploaded using
already-built Git artifacts (which is the case when Git for Windows'
"Rebase to <git-version>" PRs are "merged" by updating the `main` ref),
after verifying that there is a valid `tag-git` Check Run,
GitForWindowsHelper wants to find the associated `git-artifacts`
workflow runs.

They can be identified using the `tag-git` _workflow_ run, though, i.e.
the ID of the workflow as run in `git-for-windows-automation`, not the
mirrored _Check Run_ in `git-for-windows/git`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>