cascading-runs: do verify that the sender is trusted
The cascading runs feature of GitForWindows' GitHub App basically
listens for Check Runs to be completed, and upon their completion, other
automation is triggered. There are currently two such cascades:

- When the `tag-git` Check Run in a regular Git for Windows PR has
  completed (which is expected to be triggered via the `/git-artifacts`
  slash command by a trusted user), the corresponding `git-artifacts`
  workflow runs have been triggered, one per supported CPU architecture.

- When the `git-artifacts` Check Runs complete on a commit on
  git-for-windows/git's `main` branch that does _not_ correspond to a
  release, the `upload-snapshot` workflow needs to be triggered.

In both instances, we need to validate that the source of these Check
Runs is the intended one. We need that because Check Runs are also
created for regular workflow jobs, and so far the validation goes by
Check Run name, which is easily faked in a crafted PR.

So let's verify that the Check Run events in question were sent by a
trusted actor (and most notably _not_ by GitHub Actions implicit Check
Runs corresponding to PR workflow runs).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>